+++ test/mocks/plugins/MockAccessControlHookPlugin.sol
@@ -28,13 +28,13 @@ contract MockAccessControlHookPlugin is IValidationHook, BasePlugin {
         delete allowedTargets[msg.sender];
     }
 
-    function preUserOpValidationHook(uint8 validationId, PackedUserOperation calldata userOp, bytes32)
+    function preUserOpValidationHook(uint32 validationId, PackedUserOperation calldata userOp, bytes32)
         external
         view
         override
         returns (uint256)
     {
-        if (validationId == uint8(ValidationId.PRE_VALIDATION_HOOK)) {
+        if (validationId == uint32(ValidationId.PRE_VALIDATION_HOOK)) {
             if (bytes4(userOp.callData[:4]) == IStandardExecutor.execute.selector) {
                 address target = abi.decode(userOp.callData[4:36], (address));
 
@@ -49,13 +49,13 @@ contract MockAccessControlHookPlugin is IValidationHook, BasePlugin {
     }
 
     function preRuntimeValidationHook(
-        uint8 validationId,
+        uint32 validationId,
         address,
         uint256,
         bytes calldata data,
         bytes calldata authorization
     ) external view override {
-        if (validationId == uint8(ValidationId.PRE_VALIDATION_HOOK)) {
+        if (validationId == uint32(ValidationId.PRE_VALIDATION_HOOK)) {
             if (bytes4(data[:4]) == IStandardExecutor.execute.selector) {
                 address target = abi.decode(data[4:36], (address));
